Move IntegrityChecker into Validator target
