+++ dawn/src/dawn/Optimizer/PassIntegrityCheck.h
@@ -15,7 +15,7 @@
 #ifndef DAWN_OPTIMIZER_PASSINTEGRITYCHECK_H
 #define DAWN_OPTIMIZER_PASSINTEGRITYCHECK_H
 
-#include "dawn/Optimizer/IntegrityChecker.h"
+#include "dawn/Validator/IntegrityChecker.h"
 #include "dawn/Optimizer/Pass.h"
 
 namespace dawn {
