+++ dawn/src/dawn/Validator/IntegrityChecker.h
@@ -26,7 +26,7 @@ namespace dawn {
 //     IntegrityChecker
 //===------------------------------------------------------------------------------------------===//
 /// @brief Perform basic integrity checks on the AST.
-class IntegrityChecker : public iir::ASTVisitor {
+class IntegrityChecker : public ast::ASTVisitorForwarding {
   iir::StencilInstantiation* instantiation_;
   iir::StencilMetaInformation& metadata_;
 
@@ -45,33 +45,17 @@ public:
 
   void visit(const std::shared_ptr<iir::VarDeclStmt>& stmt) override;
 
-  void visit(const std::shared_ptr<iir::VerticalRegionDeclStmt>& stmt) override;
-
-  void visit(const std::shared_ptr<iir::StencilCallDeclStmt>& stmt) override;
-
-  void visit(const std::shared_ptr<iir::BoundaryConditionDeclStmt>& stmt) override;
-
   void visit(const std::shared_ptr<iir::AssignmentExpr>& expr) override;
 
   void visit(const std::shared_ptr<iir::UnaryOperator>& expr) override;
+
+  void visit(const std::shared_ptr<iir::ReductionOverNeighborExpr>& expr) override;
 
   void visit(const std::shared_ptr<iir::BinaryOperator>& expr) override;
 
   void visit(const std::shared_ptr<iir::TernaryOperator>& expr) override;
 
   void visit(const std::shared_ptr<iir::FunCallExpr>& expr) override;
-
-  void visit(const std::shared_ptr<iir::StencilFunCallExpr>& expr) override;
-
-  virtual void visit(const std::shared_ptr<iir::StencilFunArgExpr>& expr) override;
-
-  void visit(const std::shared_ptr<iir::VarAccessExpr>& expr) override;
-
-  void visit(const std::shared_ptr<iir::LiteralAccessExpr>& expr) override;
-
-  void visit(const std::shared_ptr<iir::FieldAccessExpr>& expr) override;
-
-  void visit(const std::shared_ptr<iir::ReductionOverNeighborExpr>& expr) override;
 
 private:
   void iterate(iir::StencilInstantiation* instantiation);
