+++ dawn/test/unit-test/dawn/CodeGen/Naive/TestCodeGenNaive.cpp
@@ -16,8 +16,8 @@
 #include "dawn/CodeGen/CodeGen.h"
 #include "dawn/Compiler/DawnCompiler.h"
 #include "dawn/Compiler/Options.h"
-#include "dawn/Optimizer/IntegrityChecker.h"
 #include "dawn/Optimizer/OptimizerContext.h"
+#include "dawn/Validator/IntegrityChecker.h"
 #include "dawn/SIR/SIR.h"
 #include "dawn/Serialization/SIRSerializer.h"
 #include "dawn/Support/DiagnosticsEngine.h"
